+++ packages/swirl-components/src/components.d.ts
@@ -1274,6 +1274,7 @@ export namespace Components {
         "min"?: number;
         "mode"?: SwirlTextInputMode1;
         "passwordToggleLabel"?: string;
+        "placeholder"?: string;
         "prefixLabel"?: string;
         "required"?: boolean;
         "rows"?: number;
@@ -3994,6 +3995,7 @@ declare namespace LocalJSX {
         "onInputFocus"?: (event: SwirlTextInputCustomEvent<FocusEvent>) => void;
         "onValueChange"?: (event: SwirlTextInputCustomEvent<string>) => void;
         "passwordToggleLabel"?: string;
+        "placeholder"?: string;
         "prefixLabel"?: string;
         "required"?: boolean;
         "rows"?: number;