+++ packages/swirl-components/src/components/swirl-checkbox/swirl-checkbox.stories.ts
@@ -9,6 +9,7 @@ export default {
     },
   },
   component: "swirl-checkbox",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,