+++ packages/swirl-components/src/components/swirl-form-control/swirl-form-control.tsx
@@ -9,6 +9,7 @@ import {
   Watch,
 } from "@stencil/core";
 import classnames from "classnames";
+import { getActiveElement } from "../../utils";
 
 export type SwirlFormControlLabelPosition = "inside" | "outside";
 
@@ -143,12 +144,14 @@ export class SwirlFormControl {
     this.hasFocus = true;
   };
 
-  private onFocusOut = () => {
-    if (!this.hasFocus) {
-      return;
+  private onKeyDown = (event: KeyboardEvent) => {
+    if (event.key === "Tab") {
+      setTimeout(() => {
+        if (!this.el.contains(getActiveElement())) {
+          this.hasFocus = false;
+        }
+      });
     }
-
-    this.hasFocus = false;
   };
 
   render() {
@@ -159,6 +162,12 @@ export class SwirlFormControl {
       ? this.inputValue.length > 0
       : Boolean(this.inputValue);
 
+    const hasCharacterCounter = Boolean(
+      this.inputEl.getAttribute("show-character-counter")
+    );
+
+    const hasPlaceholder = Boolean(this.inputEl.getAttribute("placeholder"));
+
     const isSelect = this.inputEl.tagName === "SWIRL-SELECT";
 
     const className = classnames(
@@ -166,7 +175,9 @@ export class SwirlFormControl {
       `form-control--label-position-${this.labelPosition}`,
       {
         "form-control--disabled": this.disabled,
+        "form-control--has-character-counter": hasCharacterCounter,
         "form-control--has-focus": this.hasFocus,
+        "form-control--has-placeholder": hasPlaceholder,
         "form-control--has-value": hasValue,
         "form-control--inline": this.inline,
         "form-control--invalid": this.invalid,
@@ -175,7 +186,7 @@ export class SwirlFormControl {
     );
 
     return (
-      <Host onFocusin={this.onFocusIn} onFocusout={this.onFocusOut}>
+      <Host onFocusin={this.onFocusIn} onKeyDown={this.onKeyDown}>
         <div class={className} role="group">
           <label class="form-control__label">
             <span class="form-control__label-text">{this.label}</span>