+++ packages/swirl-components/src/components/swirl-resource-list-file-item/swirl-resource-list-file-item.stories.ts
@@ -13,6 +13,7 @@ export default {
     },
   },
   component: "swirl-resource-list-file-item",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,