+++ packages/swirl-components/src/components/swirl-resource-list-item/swirl-resource-list-item.stories.ts
@@ -3,6 +3,7 @@ import Docs from "./swirl-resource-list-item.mdx";
 
 export default {
   component: "swirl-resource-list-item",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,