+++ packages/swirl-components/src/components/swirl-search/swirl-search.stories.ts
@@ -3,6 +3,7 @@ import Docs from "./swirl-search.mdx";
 
 export default {
   component: "swirl-search",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,