+++ packages/swirl-components/templates.mjs
@@ -23,7 +23,8 @@ export const cssTemplate = `:host {
 }
 `;
 
-export const docsTemplate = `import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
+export const docsTemplate = `import { Controls, Canvas, Meta, Story } from "@storybook/addon-docs";
+import * as Stories from "./{{name}}.stories";
 
 <Meta title="Components/{{pascalCase name}}" />
 
@@ -36,11 +37,9 @@ The {{pascalCase name}} component is used to …
 
 ## Usage
 
-<Canvas withSource="open">
-  <Story id="components-{{lowerCase (pascalCase name)}}--{{name}}" />
-</Canvas>
+<Canvas of={Stories.{{pascalCase name}}{{append '' '}'}} sourceState="shown"></Canvas>
 
-<ArgsTable story="." />
+<Controls of={Stories.{{pascalCase name}}{{append '' '}'}} />
 
 ## Theming
 
@@ -112,6 +111,7 @@ import Docs from "./{{name}}.mdx";
 
 export default {
   component: "{{name}}",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,